fix: restructured app, created proper folders and moved there
